[fix] Fix parameter get error
